yoav: pose renaming and design clean
