+++ src/main/java/frc/robot/Robot.java
@@ -6,8 +6,7 @@ package frc.robot;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
-import frc.robot.poseestimation.poseestimator.PoseCalculator;
-import frc.robot.poseestimation.poseestimator.PoseEstimationSuperstructure;
+import frc.robot.superstructers.poseestimator.PoseEstimatorSuperstructure;
 import frc.robot.subsystems.swerve.Swerve;
 import frc.robot.subsystems.swerve.SwerveConstants;
 import frc.robot.subsystems.swerve.SwerveState;
@@ -20,11 +19,9 @@ public class Robot {
 
     public static final RobotTypeUtils.RobotType ROBOT_TYPE = RobotTypeUtils.determineRobotType(RobotTypeUtils.RobotType.REAL);
 
-    private static final PoseCalculator poseCalculator = new PoseCalculator();
 
     public static final Swerve swerve = new Swerve();
-
-    public static final PoseEstimationSuperstructure poseEstimator = new PoseEstimationSuperstructure(poseCalculator, swerve);
+    public static final PoseEstimatorSuperstructure poseEstimator = new PoseEstimatorSuperstructure(swerve);
 
     public Robot() {
         buildPathPlannerForAuto();
