+++ src/main/java/frc/robot/subsystems/swerve/odometryThread/PhoenixOdometryThread6328.java
@@ -18,7 +18,7 @@ import com.ctre.phoenix6.StatusSignal;
 import com.ctre.phoenix6.hardware.ParentDevice;
 import com.ctre.phoenix6.jni.CANBusJNI;
 import edu.wpi.first.wpilibj.Timer;
-import frc.robot.poseestimation.poseestimator.PoseEstimatorConstants;
+import frc.robot.superstructers.poseestimator.PoseEstimatorConstants;
 import frc.robot.subsystems.swerve.Swerve;
 import frc.utils.cycletime.CycleTimeUtils;
 import org.littletonrobotics.junction.Logger;
